Fix root selection in dependency conversion
If a terminal is the head terminal of the root node in UCCA, it should
not be treated the same as an orphan node when converting to dependency,
but instead it should be marked as the root in the output.

Also, slight refactoring for consistency in head_index notation. Now it
is always 1-based, i.e. index 0 is the root.