Hack to restore acl dependencies to the right head
Due to the remote edges, there is a cycle and the topological sort
creates the acl child before the antecedent, so that there is no choice
but to attach the child to the passage root (i.e. edge.head.node is None
at the time of creation).  This is a reasonably compromise because
otherwise we would actually get a cycle in the UCCA graph.  By handling
it in the conversion back to CoNLL-U we
1. Avoid the UCCA cycle
2. Restore the dependency correctly

IT's the enhanced edge that allows us to find the correct edge, but in
treebanks without enhanced edges for relative clauses we won't even need
this solution because it will be attached correctly in the UCCA graph to
begin with, as there is no cycle in the dependency graph and the
topological sort will do a good job.